refactor: rename files and variables
